backend/requirements.txt
# Core web framework
fastapi
pydantic
taipy-gui
streamlit

# Computer vision & image processing
opencv-python
Pillow
easyocr
html2image

# Machine learning / facial recognition
deepface

# Cloud APIs
google-cloud-vision
openai
python-dotenv
requests

# Redis client
redis

# Audio / Text-to-Speech
pygame
pydub
gTTS        # (if using Google TTS fallback)

# Data handling
pandas
numpy
